+++ tybotRoute/test/conversation-iteration_test.js
@@ -1,0 +1,139 @@
+var assert = require('assert');
+let axios = require('axios');
+const tybot = require("..");
+const tybotRoute = tybot.router;
+var express = require('express');
+var app = express();
+const winston = require('../utils/winston');
+app.use("/", tybotRoute);
+app.use((err, req, res, next) => {
+  winston.error("General error", err);
+});
+require('dotenv').config();
+const bodyParser = require('body-parser');
+const { v4: uuidv4 } = require('uuid');
+const bots_data = require('./conversation-iteration_bot.js').bots_data;
+const PROJECT_ID = "projectID"; //process.env.TEST_ACTIONS_PROJECT_ID;
+const REQUEST_ID = "support-group-" + PROJECT_ID + "-" + uuidv4().replace(/-/g, "");
+const BOT_ID = "botID"; //process.env.TEST_ACTIONS_BOT_ID;
+const CHATBOT_TOKEN = "XXX"; //process.env.ACTIONS_CHATBOT_TOKEN;
+const { TiledeskChatbotUtil } = require('../utils/TiledeskChatbotUtil');
+const tilebotService = require('../services/TilebotService');
+
+let SERVER_PORT = 10001
+
+describe('Conversation for Iteration test', async () => {
+
+  let app_listener;
+  let util = new TiledeskChatbotUtil();
+
+  before(() => {
+    return new Promise(async (resolve, reject) => {
+      winston.info("Starting tilebot server...");
+      try {
+        tybot.startApp(
+          {
+            // MONGODB_URI: process.env.MONGODB_URI,
+            bots: bots_data,
+            TILEBOT_ENDPOINT: process.env.TILEBOT_ENDPOINT,
+            API_ENDPOINT: process.env.API_ENDPOINT,
+            REDIS_HOST: process.env.REDIS_HOST,
+            REDIS_PORT: process.env.REDIS_PORT,
+            REDIS_PASSWORD: process.env.REDIS_PASSWORD
+          }, () => {
+            winston.info("Tilebot route successfully started.");
+            var port = SERVER_PORT;
+            app_listener = app.listen(port, () => {
+              winston.info('Tilebot connector listening on port ' + port);
+              resolve();
+            });
+          });
+      }
+      catch (error) {
+        winston.error("error:", error)
+      }
+
+    })
+  });
+
+  after(function (done) {
+    app_listener.close(() => {
+      done();
+    });
+  });
+
+  it('iterate an iterable object', (done) => {
+
+    let listener;
+    let endpointServer = express();
+    endpointServer.use(bodyParser.json());
+    endpointServer.post('/:projectId/requests/:requestId/messages', function (req, res) {
+      res.send({ success: true });
+      const message = req.body;
+      console.log("message: ", message);
+      assert(message.attributes.commands !== null);
+      assert(message.attributes.commands.length === 2);
+      const command2 = message.attributes.commands[1];
+      assert(command2.type === "message");
+      console.log("command2: ", JSON.stringify(command2))
+      console.log("message text: ",command2.message.text)
+      assert(command2.message.text === "[object Object]");
+
+
+      util.getChatbotParameters(REQUEST_ID, (err, attributes) => {
+        if (err) {
+          assert.ok(false);
+        }
+        else {
+          assert(attributes);
+          assert.equal(typeof attributes["output"], 'object');
+          let json = attributes["output"];
+          console.log("json id: ", json.id);
+          console.log("json name: ", json.name);
+          assert(json.id === 0 || json.id === 1)
+          assert(json.name === 'John' || json.name === 'Mark')
+
+          if (json.id === 1) {
+            listener.close(() => {
+              done();
+            });
+          }
+        }
+      });
+
+    });
+
+
+    listener = endpointServer.listen(10002, '0.0.0.0', () => {
+      winston.verbose('endpointServer started' + listener.address());
+
+      const iterable = [
+        { id: 0, name: "John" },
+        { id: 1, name: "Mark" }
+      ];
+      let innerJson = JSON.stringify(iterable);
+      const text = `/iteration{'iterable_varname':${innerJson}}`;
+      console.log("text: ", text);
+
+      let request = {
+        "payload": {
+          "senderFullname": "guest#367e",
+          "type": "text",
+          "sender": "A-SENDER",
+          "recipient": REQUEST_ID,
+          "text": text,
+          "id_project": PROJECT_ID,
+          "metadata": "",
+          "request": {
+            "request_id": REQUEST_ID
+          }
+        },
+        "token": "XXX"
+      }
+      tilebotService.sendMessageToBot(request, BOT_ID, () => {
+         winston.verbose("Message sent:\n", request);
+      });
+    });
+  });
+
+});