+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Tiledesk native chatbot
 
+# 2.0.45
+- Added supporto to tags on DirAskGPTV2 and DirAddKbContent
+
+# 2.0.44
+- Added support for mcp tools list in DirAiPrompt
+
+# 2.0.43
+- Added Iteration action
+
 # 2.0.42
 - Updated AskGPTV2 with embeddings support